src/payments/pyproject.toml
[project]
name = "golden-api"
version = "0.1.0"
description = "API for interaction with the system"
readme = "README.md"

authors = [
    {name = "Neprin Team"}
]
maintainers = [
    {name = "Mikhail Neprin", email = "archemich@gmail.com"}
]

requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.116.1",
    "numpy>=2.3.2",
    "uvicorn>=0.35.0",
    "psycopg2-binary>=2.9.10"
]

[project.scripts]
golden-api = "golden.api.main:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "golden.api"
module-root = ""
namespace = true
